Refactoring of command parser
Detailed command parsing is passed to the respective Commands, rather
than done in CommandParser. Command parser does only the bare minimum
for parsing.
