+++ src/main/java/seedu/taskman/logic/commands/MarkCommand.java
@@ -1,12 +1,17 @@
 package seedu.taskman.logic.commands;
 
 import seedu.taskman.commons.exceptions.IllegalValueException;
+import seedu.taskman.logic.parser.CommandParser;
 import seedu.taskman.model.tag.Tag;
 import seedu.taskman.model.tag.UniqueTagList;
 import seedu.taskman.model.event.*;
 
 import java.util.HashSet;
 import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import static seedu.taskman.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 /**
  * Adds a Event to the task man.
@@ -23,6 +28,11 @@ public class MarkCommand extends Command {
 
     public static final String MESSAGE_SUCCESS = "New event added: %1$s";
     public static final String MESSAGE_DUPLICATE_EVENT = "This event already exists in TaskMan";
+    private static final Pattern EVENT_MARK_ARGS_FORMAT = // '/' forward slashes are reserved for delimiter prefixes
+            Pattern.compile("" + CommandParser.ArgumentPattern.TITLE
+                    + CommandParser.ArgumentPattern.SCHEDULE
+                    + CommandParser.ArgumentPattern.FREQUENCY
+                    + CommandParser.ArgumentPattern.TAG); // variable number of tags
 
     private final Event toAdd;
 
@@ -31,7 +41,7 @@ public class MarkCommand extends Command {
      *
      * @throws IllegalValueException if any of the raw values are invalid
      */
-    public MarkCommand(String title, String schedule, String frequency, Set<String> tags)
+    private MarkCommand(String title, String schedule, String frequency, Set<String> tags)
             throws IllegalValueException {
         final Set<Tag> tagSet = new HashSet<>();
         for (String tagName : tags) {
@@ -47,6 +57,24 @@ public class MarkCommand extends Command {
                     ? null
                     : new Frequency(frequency)
         );
+    }
+
+    public static Command prepareMark(String args) {
+        final Matcher matcher = EVENT_MARK_ARGS_FORMAT.matcher(args.trim());
+        // Validate arg string format
+        if (!matcher.matches()) {
+            return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, MESSAGE_USAGE));
+        }
+        try {
+            return new MarkCommand(
+                    matcher.group("title"),
+                    matcher.group("schedule"),
+                    matcher.group("frequency"),
+                    getTagsFromArgs(matcher.group("tagArguments"))
+            );
+        } catch (IllegalValueException ive) {
+            return new IncorrectCommand(ive.getMessage());
+        }
     }
 
     @Override
